% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticregression.R
\name{logisticregression}
\alias{logisticregression}
\title{Logistic Regression}
\usage{
logisticregression(formula, data)
}
\arguments{
\item{formula}{The formula of the model. Usually y ~ x1 + x2 + ...}

\item{data}{The dataset for your regression.}
}
\value{
A list of 12 contents.

1) The summary of the model, including Coefficients and their significance, and How is the model compared to the intercept-only (null) model

2) Test of each predictor of the logistic model

3) -2*LL difference:(null - model)

4) df of this chi-square test, the difference in parameter number

5) The chi-square of The overall test of the model

6) Hosmer and Lemeshow goodness of fit (GOF) test

7~9) Examining each predictor

10) Odds ratio with its confidence interval for each predictor

11) Log likelihood

12) Model fitting quality; r2CU is the Nagelkerke R2; r2ML is the Cox&Snell R2
}
\description{
Logistic regression is for fitting a regression curve to data in which the dependent variable is dichotomous.
}
