% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticregression.R
\name{logisticregression}
\alias{logisticregression}
\title{Logistic Regression}
\usage{
logisticregression(formula, data, hoslemG = 10)
}
\arguments{
\item{formula}{The formula of the model. Usually y ~ x1 + x2 + ...}

\item{data}{The dataset for your regression.}

\item{hoslemG}{g in hoslem test}
}
\value{
The summary of the model, including Coefficients and their significance, and How is the model compared to the intercept-only (null) model

Test of each predictor of the logistic model

Wald Test for intercept and each coefficient

Hosmer and Lemeshow goodness of fit (GOF) test

Odds ratio with its confidence interval for each predictor

Log likelihood

Model fitting quality: r2CU is the Nagelkerke R2; r2ML is the Cox&Snell R2

Classification Table
}
\description{
Logistic regression is for fitting a regression curve to data in which the dependent variable is dichotomous.
}
